# Use affine pedersen params