This bumps snarky as to avoid converting the Pedersen generators to affine coordinates on the fly as we prove.