@mrmr1993 2019-07-10
Combined with o1-labs/snarky#309 I'm seeing a decent improvement:
```
$ `which time` dune exec coda -- tra -witness-only -repeat 10
53.53user 1.72system 0:55.99elapsed 98%CPU (0avgtext+0avgdata 214944maxresident)k
24048inputs+8360outputs (4major+724539minor)pagefaults 0swaps
$ `which time` dune exec coda -- tra -witness-only -repeat 10
53.91user 1.38system 0:55.19elapsed 100%CPU (0avgtext+0avgdata 206996maxresident)k
0inputs+8384outputs (0major+722964minor)pagefaults 0swaps
$ (cd ..; make benchmarks); BENCHMARK_LIBRARIES=vrf_lib_tests ../scripts/benchmarks.sh
INFO Not using Docker
cd src && dune build app/benchmarks/main.exe
Running inline tests in library "vrf_lib_tests"
Estimated testing time 20s (2 benchmarks x 10s). Change using '-quota'.
┌──────────────────────────────────────────────────────────────────────┬──────────┬─────────────┬────────────┬────────────┬────────────┐
│ Name                                                                 │ Time/Run │     mWd/Run │   mjWd/Run │   Prom/Run │ Percentage │
├──────────────────────────────────────────────────────────────────────┼──────────┼─────────────┼────────────┼────────────┼────────────┤
│ [lib/vrf_lib/integrated_test.ml:vrf bench module] vrf eval unchecked │   1.25ms │     73.30kw │     4.47kw │     4.17kw │      0.28% │
│ [lib/vrf_lib/integrated_test.ml:vrf bench module] vrf eval checked   │ 439.60ms │ 20_731.45kw │ 6_899.16kw │ 6_898.35kw │    100.00% │
└──────────────────────────────────────────────────────────────────────┴──────────┴─────────────┴────────────┴────────────┴────────────┘
Benchmarks that take 1ns to 100ms can be estimated precisely. For more reliable
estimates, redesign your benchmark to have a shorter execution time.
$ (cd ..; make benchmarks); BENCHM
ARK_LIBRARIES=vrf_lib_tests ../scripts/benchmarks.sh
INFO Not using Docker
cd src && dune build app/benchmarks/main.exe
Running inline tests in library "vrf_lib_tests"
Estimated testing time 20s (2 benchmarks x 10s). Change using '-quota'.
┌──────────────────────────────────────────────────────────────────────┬──────────┬─────────────┬────────────┬────────────┬────────────┐
│ Name                                                                 │ Time/Run │     mWd/Run │   mjWd/Run │   Prom/Run │ Percentage │
├──────────────────────────────────────────────────────────────────────┼──────────┼─────────────┼────────────┼────────────┼────────────┤
│ [lib/vrf_lib/integrated_test.ml:vrf bench module] vrf eval unchecked │   1.25ms │     73.30kw │     4.47kw │     4.17kw │      0.28% │
│ [lib/vrf_lib/integrated_test.ml:vrf bench module] vrf eval checked   │ 442.93ms │ 20_731.45kw │ 6_899.16kw │ 6_898.35kw │    100.00% │
└──────────────────────────────────────────────────────────────────────┴──────────┴─────────────┴────────────┴────────────┴────────────┘
Benchmarks that take 1ns to 100ms can be estimated precisely. For more reliable
estimates, redesign your benchmark to have a shorter execution time.
```

Ballpark 9% improvement, not quite as good as we were expecting, but still pretty nice.