@imeckler 2019-07-10
I'm not sure `dune exec coda -- tra -witness-only -repeat 10` is the right thing to run. It looked to me like it was spending a lot of its time evaluating and checking the constraints which we don't do in witness generation.