@mrmr1993 2019-07-10
I can turn off `eval_constraints` and run again? Might give a better indication.