@mrmr1993 2019-07-10
This PR + o1-labs/snarky#309 (`set_eval_constraints false` in `coda.ml`)
```
$ `which time` dune exec coda -- tra -witness-only -repeat 10
33.22user 1.60system 0:34.88elapsed 99%CPU (0avgtext+0avgdata 215272maxresident)k
0inputs+8384outputs (0major+466501minor)pagefaults 0swaps
$ `which time` dune exec coda -- tra -witness-only -repeat 10
33.73user 1.58system 0:35.25elapsed 100%CPU (0avgtext+0avgdata 214360maxresident)k
0inputs+8384outputs (0major+465227minor)pagefaults 0swaps
```
o1-labs/snarky#309 alone (`set_eval_constraints false` in `coda.ml`):
```
$ `which time` dune exec coda -- tra -witness-only -repeat 10
41.78user 1.90system 0:44.40elapsed 98%CPU (0avgtext+0avgdata 212400maxresident)k
23128inputs+8384outputs (1major+488341minor)pagefaults 0swaps
$ `which time` dune exec coda -- tra -witness-only -repeat 10
40.65user 2.23system 0:42.85elapsed 100%CPU (0avgtext+0avgdata 214908maxresident)k
0inputs+8384outputs (0major+488218minor)pagefaults 0swaps
```

------------------------------------

This PR + o1-labs/snarky#309 (`set_eval_constraints false` in `coda.ml` and `run_and_check` -> `run_unchecked` in `transaction_snark.ml`):

```
$ `which time` dune exec coda -- tra -witness-only -repeat 10
30.37user 1.35system 0:32.25elapsed 98%CPU (0avgtext+0avgdata 214392maxresident)k
27456inputs+8384outputs (6major+465557minor)pagefaults 0swaps
$ `which time` dune exec coda -- tra -witness-only -repeat 10
30.48user 1.14system 0:31.55elapsed 100%CPU (0avgtext+0avgdata 214352maxresident)k
0inputs+8384outputs (0major+465910minor)pagefaults 0swaps
```

o1-labs/snarky#309 alone (`set_eval_constraints false` in `coda.ml` and `run_and_check` -> `run_unchecked` in `transaction_snark.ml`):
```
$ `which time` dune exec coda -- tra -witness-only -repeat 10
40.24user 2.22system 0:42.87elapsed 99%CPU (0avgtext+0avgdata 215716maxresident)k
22160inputs+8384outputs (4major+489263minor)pagefaults 0swaps
$ `which time` dune exec coda -- tra -witness-only -repeat 10
41.45user 1.33system 0:42.70elapsed 100%CPU (0avgtext+0avgdata 214788maxresident)k
96inputs+8384outputs (0major+489144minor)pagefaults 0swaps
```

The improvement looks even better when constraint checking is disabled :grin: